+++ src/integration-tests/breakpoints.spec.ts
@@ -22,6 +22,7 @@ import {
     gdbAsync,
     isRemoteTest,
     hardwareBreakpoint,
+    gdbNonStop,
 } from './utils';
 import { DebugProtocol } from '@vscode/debugprotocol';
 import * as os from 'os';
@@ -231,6 +232,21 @@ describe('breakpoints', async function () {
         const stoppedEvent = await stoppedEventWaitor;
         expect(stoppedEvent).to.have.property('body');
         expect(stoppedEvent.body).to.have.property('reason', 'breakpoint');
+    });
+
+    it('set an instruction breakpoint', async function () {
+        if (gdbNonStop && isRemoteTest) {
+            this.skip();
+        }
+        const bpResp = await dc.setInstructionBreakpointsRequest({
+            breakpoints: [
+                {
+                    instructionReference: '0x720',
+                },
+            ],
+        });
+        expect(bpResp.body.breakpoints.length).eq(1);
+        expect(bpResp.body.breakpoints[0].instructionReference).eq('0x720');
     });
 
     it('set type of standard breakpoint', async () => {